+++ src/main/java/ru/petrov/controller/TypeOfValueController.java
@@ -14,8 +14,6 @@ import ru.petrov.dto.TypeOfValueDto;
 import ru.petrov.model.TypeOfValue;
 import ru.petrov.service.TypeOfValueService;
 import ru.petrov.util.exception.EntityNotCreatedException;
-import ru.petrov.util.exception.EntityNotFoundException;
-import ru.petrov.util.exception.ErrorResponse;
 import ru.petrov.util.validator.TypeOfValueValidator;
 
 import java.util.List;
@@ -65,26 +63,5 @@ public class TypeOfValueController {
         Optional<TypeOfValue> save = typeService.save(type);
         return ResponseEntity.created(ServletUriComponentsBuilder.fromCurrentRequest()
                 .path("/{id}").buildAndExpand(save.get().getId()).toUri()).build();
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(code = HttpStatus.NOT_FOUND)
-    private ResponseEntity<ErrorResponse> handleException(EntityNotFoundException e) {
-        ErrorResponse response = new ErrorResponse(
-                "Тип измерения не найден",
-                System.currentTimeMillis()
-        );
-        return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(code = HttpStatus.UNPROCESSABLE_ENTITY)
-    private ResponseEntity<ErrorResponse> handleException(EntityNotCreatedException e) {
-        ErrorResponse response = new ErrorResponse(
-                e.getMessage(),
-                System.currentTimeMillis()
-        );
-        //https://stackoverflow.com/questions/16133923/400-vs-422-response-to-post-of-data
-        return new ResponseEntity<>(response, HttpStatus.UNPROCESSABLE_ENTITY);
     }
 }
